widget/wtrackmenu: move remaining actions from wtracktableview
